Canadian Health Measures Survey (CHMS) Cycle 1 Medications

This dummy data representing the medication portion of the second cycle of the Canadian Health Measures Survey (CHMS). The CHMS survey is conducted by Statistics Canada.

Usage

data(cycle1_meds)

Format

A data frame with X rows and Y columns.

Source

Statistics Canada

Examples

data(cycle1_meds)
str(cycle1_meds)
#> 'data.frame':	50 obs. of  81 variables:
#>  $ CLINICID: int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ ATC_101A: chr  "C07AA05" "C09AA06" "C08CA01" "A10BC02" ...
#>  $ ATC_102A: chr  "A01AB05" "C09AA02" "C07AB02" "C03AA03" ...
#>  $ ATC_103A: chr  "C02CC07" "C08CA06" "C07AB07" "C07AB03" ...
#>  $ ATC_104A: chr  "C03CA01" "C03CA01" "C09AA04" "C08CA02" ...
#>  $ ATC_105A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_106A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_107A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_108A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_109A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_110A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_111A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_112A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_113A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_114A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_115A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_201A: chr  "C03BA08" "C07AA07" "C07AA12" "M01AG02" ...
#>  $ ATC_202A: chr  "C07AA05" "C09AA06" "C08CA01" "A10BC02" ...
#>  $ ATC_203A: chr  "C03BA08" "C07AA07" "C07AA12" "M01AG02" ...
#>  $ ATC_204A: chr  "C03BC02" "C08CA01" "C07AA05" "C07AA06" ...
#>  $ ATC_205A: chr  "C02KX01" "C02AA05" "C07AG02" "C07AA06" ...
#>  $ ATC_206A: chr  "A01AB05" "C09AA02" "C07AB02" "C03AA03" ...
#>  $ ATC_207A: chr  "C02CC07" "C08CA06" "C07AB07" "C07AB03" ...
#>  $ ATC_208A: chr  "C03CA01" "C03CA01" "C09AA04" "C08CA02" ...
#>  $ ATC_209A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_210A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_211A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_212A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_213A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_214A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_215A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_131A: chr  "C07AA05" "C09AA06" "C08CA01" "A10BC02" ...
#>  $ ATC_132A: chr  "C03BA08" "C07AA07" "C07AA12" "M01AG02" ...
#>  $ ATC_133A: chr  "C03BC02" "C08CA01" "C07AA05" "C07AA06" ...
#>  $ ATC_134A: chr  "C02KX01" "C02AA05" "C07AG02" "C07AA06" ...
#>  $ ATC_135A: chr  "A01AB05" "C09AA02" "C07AB02" "C03AA03" ...
#>  $ ATC_231A: chr  "C02CC07" "C08CA06" "C07AB07" "C07AB03" ...
#>  $ ATC_232A: chr  "C03CA01" "C03CA01" "C09AA04" "C08CA02" ...
#>  $ ATC_233A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_234A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ ATC_235A: chr  "C09AA04" "C07AB02" "C08CA02" "A10BD05" ...
#>  $ MHR_101B: num  2 3 1 3 2 3 1 3 2 3 ...
#>  $ MHR_102B: num  1 4 3 2 1 4 3 2 1 4 ...
#>  $ MHR_103B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_104B: num  1 2 4 3 1 2 4 3 1 2 ...
#>  $ MHR_105B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_106B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_107B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_108B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_109B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_110B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_111B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_112B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_113B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_114B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_115B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_131B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_132B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_133B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_134B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_135B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_201B: num  3 2 2 3 3 2 2 3 3 2 ...
#>  $ MHR_202B: num  2 3 1 3 2 3 1 3 2 3 ...
#>  $ MHR_203B: num  3 2 2 3 3 2 2 3 3 2 ...
#>  $ MHR_204B: num  3 1 5 1 3 1 5 1 3 1 ...
#>  $ MHR_205B: num  1 2 3 4 1 2 3 4 1 2 ...
#>  $ MHR_206B: num  1 4 3 2 1 4 3 2 1 4 ...
#>  $ MHR_207B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_208B: num  1 2 4 3 1 2 4 3 1 2 ...
#>  $ MHR_209B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_210B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_211B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_212B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_213B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_214B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_215B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_231B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_232B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_233B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_234B: num  3 2 1 4 3 2 1 4 3 2 ...
#>  $ MHR_235B: num  3 2 1 4 3 2 1 4 3 2 ...