CHMS Cycle 3 Medications

This synthetic data representing the medication portion of the third cycle of the Canadian Health Measures Survey (CHMS). The CHMS survey is conducted by Statistics Canada.

Value

cycle3_meds

a data frame

Examples

data(cycle3_meds)
str(cycle3_meds)
#> 'data.frame':	48 obs. of  3 variables:
#>  $ clinicid: num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ meucatc : chr  "C07AA05" "C09AA06" "C08CA01" "A10BC02" ...
#>  $ npi_25b : num  2 3 1 3 2 3 1 3 2 3 ...