Non-steroidal anti-inflammatory drugs (NSAIDs)

This function checks if a given medication is an NSAID. This function processes multiple inputs efficiently.

Usage

is_NSAID(MEUCATC, NPI_25B)

Arguments

MEUCATC

character ATC code of the medication.

NPI_25B

integer Time when the medication was last taken.

Value

numeric 1 if medication is an NSAID, 0 otherwise. If inputs are invalid or out of bounds, the function returns a tagged NA.

Details

Identifies NSAIDs based on ATC codes starting with "M01A".

     **Missing Data Codes:**
     - `MEUCATC`: `9999996` (Not applicable), `9999997-9999999` (Missing)
     - `NPI_25B`: `6` (Not applicable), `7-9` (Missing)

Examples

# Scalar usage: Single respondent
is_NSAID("M01AB05", 1)
#> [1] 1
# Returns: 1

# Example: Respondent has non-response values for all inputs.
result <- is_NSAID("9999998", 8)
result # Shows: NA
#> [1] NA
haven::is_tagged_na(result, "b") # Shows: TRUE (confirms it's tagged NA(b))
#> [1] TRUE
format(result, tag = TRUE) # Shows: "NA(b)" (displays the tag)
#> [1] "NA"

# Multiple respondents
is_NSAID(c("M01AB05", "A10BB09"), c(1, 3))
#> [1] 1 0
# Returns: c(1, 0)