variable-details.csv

This dataset provides details on how variables are recoded in chmsflow.

Value

variable_details

a data frame

Details

See the below link for more details about how the worksheet is structured https://big-life-lab.github.io/chmsflow/articles/variable_details.html

Examples

data(variable_details)
str(variable_details)
#> 'data.frame':	1074 obs. of  16 variables:
#>  $ variable               : chr  "acemed" "acemed" "acemed" "acemed" ...
#>  $ dummyVariable          : chr  "N/A" "acemed_cat2_1" "acemed_cat2_2" "acemed_cat2_NA::b" ...
#>  $ typeEnd                : chr  "cat" "cat" "cat" "cat" ...
#>  $ databaseStart          : chr  "cycle1_meds, cycle2_meds" "cycle1_meds, cycle2_meds" "cycle1_meds, cycle2_meds" "cycle1_meds, cycle2_meds" ...
#>  $ variableStart          : chr  "DerivedVar::[atc_101a, atc_102a, atc_103a, atc_104a, atc_105a, atc_106a, atc_107a, atc_108a, atc_109a, atc_110a"| __truncated__ "DerivedVar::[atc_101a, atc_102a, atc_103a, atc_104a, atc_105a, atc_106a, atc_107a, atc_108a, atc_109a, atc_110a"| __truncated__ "DerivedVar::[atc_101a, atc_102a, atc_103a, atc_104a, atc_105a, atc_106a, atc_107a, atc_108a, atc_109a, atc_110a"| __truncated__ "DerivedVar::[atc_101a, atc_102a, atc_103a, atc_104a, atc_105a, atc_106a, atc_107a, atc_108a, atc_109a, atc_110a"| __truncated__ ...
#>  $ typeStart              : chr  "N/A" "N/A" "N/A" "N/A" ...
#>  $ recEnd                 : chr  "Func::cycles1to2_ace_inhibitors" "1" "0" "NA::b" ...
#>  $ numValidCat            : chr  "N/A" "2" "2" "2" ...
#>  $ catLabel               : chr  "N/A" "Yes" "No" "missing" ...
#>  $ catLabelLong           : chr  "N/A" "Yes" "No" "missing" ...
#>  $ units                  : chr  "N/A" "N/A" "N/A" "N/A" ...
#>  $ recStart               : chr  "N/A" "N/A" "N/A" "N/A" ...
#>  $ catStartLabel          : chr  "N/A" "N/A" "N/A" "N/A" ...
#>  $ variableStartShortLabel: chr  "ACE inhibitors" "ACE inhibitors" "ACE inhibitors" "ACE inhibitors" ...
#>  $ variableStartLabel     : chr  "Taking ACE inhibitors" "Taking ACE inhibitors" "Taking ACE inhibitors" "Taking ACE inhibitors" ...
#>  $ notes                  : chr  "" "" "" "" ...